+++ src/TimestampableImmutable/UpdatedAt.php
@@ -19,7 +19,7 @@ trait UpdatedAt
      * @ORM\Column(type="datetime", name="updated_at", nullable=true)
      */
     #[Column(type: "datetime", name: "updated_at", nullable: true)]
-    private ?\DateTimeImmutable $updatedAt;
+    protected ?\DateTimeImmutable $updatedAt;
 
 
     // Accessors :